# planning/openapi.yaml
openapi: 3.1.0
info:
  title: EspaciGo API
  version: 1.0.0
  description: 'Prototipo local sintético. Reseñas y avisos M09 usan la decisión D-COMM local; no son servicios productivos ni plazos legales.'
servers:
- url: /api/v1
  description: Base relativa versionada; el host se configura por entorno.
security:
- bearerAuth: []
paths:
...
  /auth/session:
    get:
      operationId: currentSession
      tags:
      - M01 local
      description: Consulta solo la sesión del Bearer actual y sus roles/estado; no renueva inactividad. No
        acepta IDs de terceros ni parámetros de URL.
      responses:
        '400': *id001
        '403': *id002
        '404': *id003
        '405': *id004
        '415': *id005
        '422': *id006
        '429': *id007
        '500': *id008
        '503': *id009
        '200':
          description: Operación completada.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /local/booking-trial/reservations:
    get:
      operationId: listOwnLocalTrialReservations
      tags: [M06 local]
      description: Historial de reservas del participante autenticado en el fixture privado.
      responses:
        '200': { description: Historial y aviso ENSAYO LOCAL — SIN COBRO REAL., content: { application/json: { schema: { $ref: '#/components/schemas/LocalBookingListEnvelope' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: requestLocalTrialReservation
      description: Revalida horario, ocupación y elegibilidad KYC sintética de anfitrión y arrendatario dentro de la transacción. La falta de KYC, una cotización que ya no cabe u ocupación incompatible producen conflicto sin crear reserva ni retención. El cambio de elegibilidad no modifica reservas ni historiales previos.
      tags: [M06 local]
      parameters:
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 1, maxLength: 200 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quote_id]
              properties:
                quote_id: { type: string, format: uuid }
      responses:
        '201': { description: Retención en ocupacion y reserva pendiente de pago dentro de una transacción., content: { application/json: { schema: { $ref: '#/components/schemas/LocalBookingReservationEnvelope' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /auth/logout:
    post:
      operationId: logout
      tags:
      - M01 local
      description: 'CU-06/RQF-023: requiere sesión válida; revoca credencial. Replay/expirada devuelve 401.'
      responses:
        '400': *id001
        '403': *id002
        '404': *id003
        '405': *id004
        '415': *id005
        '422': *id006
        '429': *id007
        '500': *id008
        '503': *id009
        '204':
          description: Operación completada.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Credencial de acceso protegida por TLS. Esta base no fija formato, emisor ni ciclo de vida
        del token; cada operación declara sus excepciones.
  headers:
    X-Request-ID:
      description: Identificador de correlación generado o propagado por el servidor.
      schema:
        type: string
        format: uuid
    WWW-Authenticate:
      description: Desafío Bearer que acompaña una respuesta 401.
      schema:
        type: string
      example: Bearer
...
  schemas:
...
    LocalBookingReservationEnvelope:
      type: object
      required: [data, safety_notice]
      properties:
        safety_notice: { type: string, const: "ENSAYO LOCAL — SIN COBRO REAL" }
        data: { $ref: '#/components/schemas/LocalBookingReservation' }
...
    LocalBookingListEnvelope:
      type: object
      required: [data, safety_notice]
      properties:
        safety_notice: { type: string, const: "ENSAYO LOCAL — SIN COBRO REAL" }
        data:
          type: object
          required: [items]
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/LocalBookingReservation' } }
...
    LocalBookingReservation:
      type: object
      required: [id, quote_id, space_id, host_id, renter_id, state, rate_unit, unit_price_clp, currency, units, subtotal_clp, start_at, end_at, time_zone, conditions, cancellation_policy_version, pay_expires_at, created_at, updated_at, unread_count]
      properties:
        id: { type: string, format: uuid }
        quote_id: { type: string, format: uuid }
        space_id: { type: string, format: uuid }
        host_id: { type: string, format: uuid }
        renter_id: { type: string, format: uuid }
        state: { type: string, enum: [pendiente_de_pago, pagada, aprobada_host, firma_parcial, lista_para_checkin, en_curso, finalizada, en_disputa, cancelada_por_firma, cancelada_por_pago, rechazada_arrendador, vencida_pago, vencida_host, cancelada_arrendatario] }
        rate_unit: { type: string, enum: [hora, dia, mes] }
        unit_price_clp: { type: integer }
        currency: { type: string, const: CLP }
        units: { type: integer }
        subtotal_clp: { type: integer }
        start_at: { type: string, format: date-time }
        end_at: { type: string, format: date-time }
        time_zone: { type: string }
        conditions: { type: string }
        cancellation_policy_version: { type: string, const: local_flexible_v1, description: Snapshot local tomado desde la cotización. }
        guarantee_policy_version: { type: [string, 'null'], enum: [garantia_local_fija_v1, null] }
        guarantee_currency: { type: [string, 'null'], enum: [CLP, null] }
        guarantee_expected_clp: { type: [integer, 'null'], enum: [50000, null], description: Ausente/null únicamente para reservas históricas previas a V40. }
        refund_id: { type: string, format: uuid }
        refund_operation_id: { type: string, format: uuid }
        refund_amount_clp: { type: integer, minimum: 1 }
        refund_state: { type: string, enum: [pendiente, completada] }
        refund_last_result: { type: string, enum: [exito_simulado, fallo_simulado, sin_respuesta_simulada] }
        refund_updated_at: { type: string, format: date-time }
        pay_expires_at: { type: string, format: date-time }
        host_expires_at: { type: [string, 'null'], format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        unread_count: { type: integer, minimum: 0 }
...
    Error:
      type: object
      additionalProperties: false
      required:
      - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      additionalProperties: false
      required:
      - code
      - message
      - request_id
      properties:
        code:
          type: string
          description: Código estable, legible por máquina y documentado por operación.
          example: validation_error
        message:
          type: string
          description: Mensaje seguro y no sensible para presentación o diagnóstico.
          example: La solicitud contiene campos inválidos.
        request_id:
          type: string
          format: uuid
        details:
          type: array
          description: Detalles seguros, opcionales y específicos del error.
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      additionalProperties: false
      required:
      - field
      - code
      properties:
        field:
          type: string
        code:
          type: string
        message:
          type: string
...
    SessionResponse:
      type: object
      additionalProperties: false
      required:
      - account_id
      - roles
      - restricted_mode
      properties:
        account_id: *id010
        roles:
          type: array
          items:
            type: string
            enum:
            - arrendatario
            - arrendador
            - administrador
        restricted_mode:
          type: boolean
          description: La cuenta sigue bloqueada administrativamente y requiere operaciones de reserva existentes.
...
  responses:
...
    Unauthorized:
      description: Credencial ausente, inválida o expirada; se envía un desafío aplicable.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/X-Request-ID'
        WWW-Authenticate:
          $ref: '#/components/headers/WWW-Authenticate'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            unauthenticated:
              value:
                error:
                  code: unauthenticated
                  message: Se requiere autenticación.
                  request_id: 550e8400-e29b-41d4-a716-446655440000
...
    NotFound:
      description: Recurso inexistente o no visible para esta identidad.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/X-Request-ID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflicto con el estado actual, versión o clave de idempotencia.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/X-Request-ID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            conflict:
              value:
                error:
                  code: conflict
                  message: La operación no es compatible con el estado actual.
                  request_id: 550e8400-e29b-41d4-a716-446655440000
...